version control added

+++ web/edit/modules/files.py
@@ -29,16 +29,65 @@ import core.acl as acl
 from utils.utils import getMimeType
 from utils.fileutils import importFile, getImportDir, importFileIntoDir
 from contenttypes.image import makeThumbNail, makePresentationFormat
-
+from core.translation import lang, t
 
 def getContent(req, ids):
+    ret = ""
     user = users.getUserFromRequest(req)
     node = tree.getNode(ids[0])
     update_error = False
     access = acl.AccessData(req)
     if not access.hasWriteAccess(node) or "files" in users.getHideMenusForUser(user):
         return req.getTAL("web/edit/edit.html", {}, macro="access_error")
-    
+
+    masklist = []
+
+    for m in node.getType().getMasks(type="edit"):
+        if access.hasReadAccess(m):
+            masklist.append(m)
+            
+    if hasattr(node, "metaFields"):
+                
+        class SystemMask:
+            def __init__(self, name, description, fields):
+                self.name,self.description,self.fields = name,description,fields
+            def getName(self):
+                return self.name
+            def getDescription(self):
+                return self.description
+            def getDefaultMask(self):
+                return False
+            def metaFields(self, lang=None):
+                return self.fields
+            def i_am_not_a_mask():
+                pass
+        masklist = [SystemMask("settings", t(req, "settings"), node.metaFields(lang(req)))] + masklist
+
+    default = None
+    for m in masklist:
+        if m.getDefaultMask():
+            default = m
+            break
+    if not default and len(masklist):
+        default = masklist[0]
+
+    maskname = req.params.get("mask", node.get("edit.lastmask") or "editmask")
+    if maskname=="":
+        maskname = default.getName()
+
+    mask = None
+    for m in masklist:
+        if maskname==m.getName():
+            mask = m
+            break
+
+    if not mask and default:
+        mask = default
+        maskname = default.getName()
+
+    if not mask:
+        return req.getTAL("web/edit/modules/metadata.html", {}, macro="no_mask")
+
     if 'data' in req.params:
         if req.params.get('data')=='children': # get formated list of childnodes of selected directory
             req.writeTAL("web/edit/modules/files.html", {'children':node.getChildren()}, macro="edit_files_popup_children")
@@ -124,14 +173,35 @@ def getContent(req, ids):
             uploadfile = req.params.get("updatefile")
 
             if uploadfile:
-                if req.params.get("change_file")=="yes": # remove old files
+                create_version_error = False
+                # Create new version when change file
+                if (req.params.get('generate_new_version') and not hasattr(node, "metaFields")):
+                    if (req.params.get('version_comment', '').strip()==''
+                        or req.params.get('version_comment', '').strip()=='&nbsp;'):
+                        create_version_error = True
+                        ret += req.getTAL("web/edit/modules/files.html", {}, macro="version_error")
+                    else:
+                        current = node
+                        node = node.createNewVersion(user)
+                        
+                        for item in mask.getMaskFields():
+                            field = item.getField()
+                            if field and field.getContentType()=="metafield":
+                                node.set(field.getName(), current.get(field.getName()))
+                        ret += req.getTAL("web/edit/modules/metadata.html", {'url':'?id='+node.id+'&tab=files', 'pid':None}, macro="redirect")
+
+                if req.params.get("change_file")=="yes" and not create_version_error: # remove old files
                     for f in node.getFiles():
                         if f.getType() in node.getSysFiles():
                             node.removeFile(f)
                             if os.path.exists(f.retrieveFile()): # delete file from disc
                                 os.remove(f.retrieveFile())
+                    node.set("system.version.comment", '('+t(req, "edit_files_new_version_exchanging_comment")+')\n'+req.params.get('version_comment', ''))
 
-                if req.params.get("change_file") in ["yes", "no"]:
+                if req.params.get("change_file")=="no" and not create_version_error:
+                    node.set("system.version.comment", '('+t(req, "edit_files_new_version_adding_comment")+')\n'+req.params.get('version_comment', ''))
+
+                if req.params.get("change_file") in ["yes", "no"] and not create_version_error:
                     file = importFile(uploadfile.filename, uploadfile.tempname) # add new file
                     node.addFile(file)
                     logging.getLogger('usertracing').info(user.name+" changed file of node "+node.id+" to "+uploadfile.filename+" ("+uploadfile.tempname+")")
@@ -142,7 +212,7 @@ def getContent(req, ids):
                         attpath = f.getName()
                         break
 
-                if req.params.get("change_file")=="attdir": # add attachmentdir
+                if req.params.get("change_file")=="attdir" and not create_version_error: # add attachmentdir
                     dirname = req.params.get("inputname")
 
                     if attpath=="": # add attachment directory
@@ -152,10 +222,11 @@ def getContent(req, ids):
                             node.addFile(tree.FileNode(name=getImportDir() + "/" + attpath, mimetype="inode/directory", type="attachment"))
 
                         file = importFileIntoDir(getImportDir() + "/" + attpath, uploadfile.tempname) # add new file
+                    node.set("system.version.comment", '('+t(req, "edit_files_new_version_attachment_directory_comment")+')\n'+req.params.get('version_comment', ''))
                     pass
                 
                 
-                if req.params.get("change_file")=="attfile": # add file as attachment
+                if req.params.get("change_file")=="attfile" and not create_version_error: # add file as attachment
                     if attpath=="":
                         # no attachment directory existing
                         file = importFile(uploadfile.filename, uploadfile.tempname) # add new file
@@ -165,6 +236,7 @@ def getContent(req, ids):
                     else:
                         # import attachment file into existing attachment directory
                         file = importFileIntoDir(getImportDir() + "/" + attpath, uploadfile.tempname) # add new file
+                    node.set("system.version.comment", '('+t(req, "edit_files_new_version_attachment_comment")+')\n'+req.params.get('version_comment', ''))
                     pass
 
             
@@ -205,7 +277,8 @@ def getContent(req, ids):
     v["node"] = node
     v["update_error"] = update_error
     v["user"] = user
-    
+    v["access"] = access
+
     v["files"] = filter(lambda x: x.type!='statistic', node.getFiles())
     v["statfiles"] = filter(lambda x: x.type=='statistic', node.getFiles())
     v["attfiles"] = filter(lambda x: x.type=='attachment', node.getFiles())
@@ -217,5 +290,6 @@ def getContent(req, ids):
                     af = tree.FileNode(root+"/"+name, "attachmentfile", getMimeType(name)[0])
                     v["att"].append(af)
 
-    return req.getTAL("web/edit/modules/files.html", v, macro="edit_files_file")
+    ret += req.getTAL("web/edit/modules/files.html", v, macro="edit_files_file")
+    return ret
